Add files via upload
In this project, I proficiently performed data cleaning and preprocessing using Python's NumPy and Pandas libraries, ensuring data quality and readiness for analysis. The core of the project involved conducting a comprehensive Exploratory Data Analysis (EDA) on the Zomato Chennai dataset. I employed data visualization libraries such as Matplotlib, Seaborn, and Plotly to visualize and analyze the data, facilitating a deeper understanding. This included one-variable analysis to reveal kurtosis and skewness insights and two-variable analysis to uncover valuable patterns and relationships. The results of this EDA serve as a foundation for data-driven decision-making.